Process-API 04: the exec family
The 'l' in function indicates the args are passed
as variadic arguments terminated by a NULL pointer.

The 'v' indicates the args are passed with a vector,
whose last element should be NULL pointer.

The 'p' indicates we can pass only the filename,
it will look up this file for us with the `PATH` env.

The 'e' indicates we can pass a customized envs,
when not specified, it is the extern char **environ.

Signed-off-by: Qi Wang <mpiglet@outlook.com>